LayoutTests/platform/ios-simulator/editing/pasteboard/paste-4039777-fix-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldInsertNode:#document-fragment replacingDOMRange:range from 0 of #text > LI > UL > DIV > DIV > BODY > HTML > #document to 14 of #text > LI > UL > DIV > DIV > DIV > BODY > HTML > #document givenAction:WebViewInsertActionPasted
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 2 of UL > DIV > DIV > BODY > HTML > #document to 0 of UL > DIV > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > LI > UL > DIV > DIV > BODY > HTML > #document to 1 of #text > LI > UL > DIV > UL > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: shouldApplyStyle:<DOMCSSStyleDeclaration: 0x7fc61c894ac0> toElementsInDOMRange:range from 0 of #text > LI > UL > DIV > DIV > BODY > HTML > #document to 1 of #text > LI > UL > DIV > UL > DIV > DIV > BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 1 of LI > UL > DIV > DIV > BODY > HTML > #document to 1 of #text > SPAN > FONT > LI > UL > DIV > UL > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > SPAN > FONT > LI > UL > DIV > DIV > BODY > HTML > #document to 1 of #text > SPAN > FONT > LI > UL > DIV > UL > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
Bug fix for <a href="rdar://problem/4039777">&lt;rdar://problem/4039777&gt;</a> Pasting particular snippet of HTML containing list items and a link creates too many list items

Expected result:
| "
"
| <ul>
|   style="text-align:right;"
|   <li>
|     "A"
| "
"
| <div>
|   "
    "
|   <ul>
|     "
        "
|     <li>
|       "
            "
|       <a>
|         href=""
|         "B"
|       "
            "
|       <br>
|       "
            C 
        "
|     "
    "
|   "
"
| "
"

Actual result:
| "
"
| <ul>
|   style="text-align:right;"
|   <li>
|     style="text-align: start;"
|     <font>
|       size="3"
|       <span>
|         style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|         "A"
|   <div>
|     style="text-align: start;"
|     <ul>
|       <li>
|         <font>
|           color="#000000"
|           size="3"
|           <span>
|             style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|             <a>
|               href=""
|               "B"
|             " "
|             <br>
|             "C<#selection-caret>"
| <div>
|   <ul>
|     "
    "
|   "
"
| "
"
